requirements.txt
streamlit
numpy
pillow
torch==2.0.1
torchvision==0.15.2
